# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  database.Partner:
    properties:
      companyName:
        type: string
      createdAt:
        type: string
      id:
        type: string
      metaData:
        additionalProperties: true
        type: object
      parent:
        $ref: '#/definitions/database.Partner'
      parentId:
        type: string
      referralCode:
        type: string
      socialMedia:
        additionalProperties: true
        description: |-
          SocialMedia mirrors PartnerApplication.SocialMedia at approve time.
          Application stays as the review snapshot; this column is the
          partner's *current* profile (Settings edits land here).

          `serializer:json` round-trips `map[string]interface{}` through
          encoding/json so the column can stay TEXT — works with both lib/pq
          and pgx. Partial `Updates(map)` bypass the serializer, so the
          service marshals the value to a JSON string before assigning.
        type: object
      status:
        $ref: '#/definitions/database.PartnerStatus'
      tier:
        $ref: '#/definitions/database.PartnerTier'
      tierId:
        type: string
      totalEarnings:
        type: number
      totalReferrals:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/database.User'
      userId:
        type: string
      website:
        type: string
    type: object
...
  database.PartnerStatus:
    enum:
    - pending
    - active
    - suspended
    type: string
    x-enum-varnames:
    - PartnerStatusPending
    - PartnerStatusActive
    - PartnerStatusSuspended
  database.PartnerTier:
    properties:
      color:
        type: string
      commissionRate:
        type: number
      createdAt:
        type: string
      id:
        type: string
      isDefault:
        type: boolean
      level:
        type: integer
      maxVolume:
        type: number
      minActiveClients:
        type: integer
      minVolume:
        type: number
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  database.SubAffiliateInvite:
    properties:
      acceptedAt:
        type: string
      createdAt:
        type: string
      email:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      inviteCode:
        type: string
      overrideRate:
        type: number
      partner:
        $ref: '#/definitions/database.Partner'
      partnerId:
        type: string
      status:
        type: string
      subPartnerId:
        type: string
      updatedAt:
        type: string
    type: object
  database.User:
    properties:
      binanceEmail:
        type: string
      canTrade:
        type: boolean
      canWithdraw:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isWhitelistEnabled:
        type: boolean
      kycLevel:
        type: integer
      lastName:
        type: string
      metaData:
        additionalProperties: true
        type: object
      status:
        type: integer
      subAccountId:
        type: string
      updatedAt:
        type: string
      vipLevel:
        type: integer
    type: object
...
  structs.ErrorResponse:
    properties:
      body:
        type: string
      error_msg:
        type: string
      status_code:
        type: integer
    type: object
...
  structs.PaginationResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
  structs.PartnerListParams:
    properties:
      current:
        type: integer
      export:
        type: boolean
      limit:
        description: alias for PageSize
        type: integer
      page:
        description: alias for Current
        type: integer
      pageSize:
        type: integer
      query:
        type: string
      sortDate:
        $ref: '#/definitions/structs.SortDate'
      status:
        type: string
      tierId:
        type: string
    type: object
...
  structs.ResponseBody:
    properties:
      body: {}
      message:
        type: string
    type: object
  structs.SortDate:
    properties:
      end_day:
        type: string
      start_day:
        type: string
    type: object
  structs.SubAffiliateInviteParams:
    properties:
      email:
        type: string
      overrideRate:
        type: number
    required:
    - email
    type: object
  structs.SubAffiliateListParams:
    properties:
      current:
        type: integer
      export:
        type: boolean
      limit:
        description: alias for PageSize
        type: integer
      page:
        description: alias for Current
        type: integer
      pageSize:
        type: integer
      sortDate:
        $ref: '#/definitions/structs.SortDate'
      status:
        type: string
    type: object
  structs.SubAffiliateStatsParams:
    properties:
      current:
        type: integer
      export:
        type: boolean
      limit:
        description: alias for PageSize
        type: integer
      page:
        description: alias for Current
        type: integer
      pageSize:
        type: integer
      sortDate:
        $ref: '#/definitions/structs.SortDate'
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: |-
    Partner program backend — public, partner, admin, system endpoints.
    Authentication is split into three layers:
    • BearerAuth (PartnerAuth/AdminAuth): Cognito ID token in `Authorization: Bearer …`.
    • InternalKey: `X-Internal-Key` header — only the xmeta-monorepo server should send this.
  title: xmeta Partner API
  version: "1.0"
paths:
...
  /admin/partner/partners/list:
    post:
      consumes:
      - application/json
      description: Returns a paginated list of partners for admin management
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PartnerListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List partners
      tags:
      - Admin Partners
...
  /partner/sub-affiliates/invite:
    post:
      consumes:
      - application/json
      description: Creates a new sub-affiliate invite under the authenticated partner
      parameters:
      - description: Invite payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.SubAffiliateInviteParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.SubAffiliateInvite'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite a sub-affiliate
      tags:
      - Partner Sub-Affiliates
  /partner/sub-affiliates/invite/{code}:
    get:
      consumes:
      - application/json
      description: Returns the sub-affiliate invite identified by code
      parameters:
      - description: Invite code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/database.SubAffiliateInvite'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get sub-affiliate invite detail
      tags:
      - Partner Sub-Affiliates
  /partner/sub-affiliates/list:
    post:
      consumes:
      - application/json
      description: Returns a paginated list of sub-affiliates under the partner
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.SubAffiliateListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                body:
                  $ref: '#/definitions/structs.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List sub-affiliates
      tags:
      - Partner Sub-Affiliates
  /partner/sub-affiliates/stats:
    post:
      consumes:
      - application/json
      description: Returns aggregate statistics for the partner's sub-affiliate network
      parameters:
      - description: Stats filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.SubAffiliateStatsParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Sub-affiliate stats
      tags:
      - Partner Sub-Affiliates
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '"Bearer {Cognito ID Token}" format'
    in: header
    name: Authorization
    type: apiKey
  InternalKey:
    description: Server-to-server API key shared with xmeta-monorepo
    in: header
    name: X-Internal-Key
    type: apiKey
swagger: "2.0"